pdf-outline/requirements.txt
fitz==0.0.1.dev2
paddleocr==2.8.0
paddleocr.egg==info
pdf2image==1.17.0
torch==2.1.0+cpu
transformers==4.53.2
PyPDF2==3.0.1
pdfminer.six==20221105
pdfplumber==0.10.2
Pillow==10.0.0
pytesseract==0.3.10